mchaib update 2016-12-14-Food-Data-Expedition.md
Added new dates for Food Hackadays in Lausanne and Basel
Initiated the idea of the online Food Data Expedition
Added Data Packagist to the list of things we want to do/use in the Expedition
Added my name as a facilitator (but no link to the school of data users yet)
Updated most links. Did not provide an end date to the online expedition yet.